Omstrukturere integrasjonstester soeknad
